Bring PoC to latest V3.23.1 at revision 3591

+++ src/corelib/Util.cpp
@@ -238,10 +238,10 @@ size_t _writecbc(FILE *fp, const unsigned char* buffer, int length, unsigned cha
 * and -1 is returned if it matches. (used in V3)
 */
 size_t _readcbc(FILE *fp,
-                 unsigned char* &buffer, unsigned int &buffer_len, unsigned char &type,
-                 Fish *Algorithm, unsigned char* cbcbuffer,
-                 const unsigned char *TERMINAL_BLOCK, size_t file_len, bool bSkip,
-                 unsigned char *pSkipTypes)
+                unsigned char* &buffer, size_t &buffer_len, unsigned char &type,
+                Fish *Algorithm, unsigned char* cbcbuffer,
+                const unsigned char *TERMINAL_BLOCK, size_t file_len, bool bSkip,
+                unsigned char *pSkipTypes)
 {
   const unsigned int BS = Algorithm->GetBlockSize();
   size_t num_bytes_read = 0;
